Mass move and comment out of useless events
Many of the german game events are too unique for us to paint in a colour of fallout. However they hold coding tricks we can learn from. Until we are comfortable at coding we will keep these on hand to look at for ideas but they will no longer be accessable via gameplay.

Co-Authored-By: SectionXIII <57029119+SectionXIII@users.noreply.github.com>