Renaming the paltry bandit raid to Mojave Bandit raid
The purpose of this event refluffing is to help tell the story of the game in a more logical manner. Random bandits = boring. Mojave bandits = evil and a interest for both NPCs and Players. + I am pretty sure the bandits down south and in the north aren't as big of a threat as the ones in the Mojave. I guess the drinking water is just diffrent over there